Code reorganization.
* Moves all the todoist api specific code into a subpackage.
* Adds Makefile to handle build, release and clean tasks.
* Adds a bit of documentation on the api package methods.
