+++ .gitignore
@@ -1,3 +1,5 @@
+release/
+
 # Compiled Object files, Static and Dynamic libs (Shared Objects)
 *.o
 *.a
